Add dir exist checks

+++ monkey_web/recipes/rm.rb
@@ -1,4 +1,5 @@
 execute 'remove containers' do
+    not_if { Dir.exist?("/srv/monkey_web/") }
     cwd '/srv/monkey_web/'
     command 'docker-compose rm --force'
     case node[:platform]
